requirements.txt
fastapi
uvicorn
numpy
pillow
tensorflow
torch
gunicorn
tensorflow-gpu==2.8.0
opencv-python
pandas